requirements.txt
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
configobj==5.0.6
configparser==5.3.0
etelemetry==0.3.0
filelock==3.8.0
fitz==0.0.1.dev2
future==0.18.2
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
isodate==0.6.1
looseversion==1.0.1
lxml==4.9.1
networkx==2.8.6
nibabel==4.0.1
nipype==1.8.3
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathlib==1.0.1
protobuf==4.21.5
prov==2.0.0
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.2
pygsheets==2.0.5
pyparsing==3.0.9
PyPDF2==2.10.3
python-dateutil==2.8.2
pytz==2022.2.1
pyxnat==1.4
rdflib==6.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.40
traits==6.4.1
uritemplate==4.1.1
urllib3==1.26.12